BackEnd/requirements.txt
# Lista de dependencias para el backend
# Ejemplo:
# flask==2.3.4
Django>=4.2
mysqlclient
djangorestframework
python-dotenv
djangorestframework-simplejwt
django-cors-headers
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-storages[s3]==1.14.2
boto3==1.34.0